+++ src/main/java/gov/nasa/pds/registry/mgr/cmd/CreateRegistryCmd.java
@@ -9,6 +9,7 @@ import org.elasticsearch.client.ResponseException;
 import org.elasticsearch.client.RestClient;
 
 import gov.nasa.pds.registry.mgr.Constants;
+import gov.nasa.pds.registry.mgr.es.client.EsClientFactory;
 import gov.nasa.pds.registry.mgr.util.CloseUtils;
 import gov.nasa.pds.registry.mgr.util.es.EsRequestBuilder;
 import gov.nasa.pds.registry.mgr.util.es.EsUtils;
@@ -46,7 +47,7 @@ public class CreateRegistryCmd implements CliCommand
         System.out.println("Elasticsearch URL: " + esUrl);
         System.out.println();
         
-        client = EsUtils.createClient(esUrl, authPath);
+        client = EsClientFactory.createRestClient(esUrl, authPath);
 
         try
         {
