+++ src/main/java/gov/nasa/pds/registry/mgr/cmd/SetArchiveStatusCmd.java
@@ -16,6 +16,7 @@ import org.elasticsearch.client.RestClient;
 import com.google.gson.Gson;
 
 import gov.nasa.pds.registry.mgr.Constants;
+import gov.nasa.pds.registry.mgr.es.client.EsClientFactory;
 import gov.nasa.pds.registry.mgr.util.CloseUtils;
 import gov.nasa.pds.registry.mgr.util.es.EsRequestBuilder;
 import gov.nasa.pds.registry.mgr.util.es.EsUtils;
@@ -81,7 +82,7 @@ public class SetArchiveStatusCmd implements CliCommand
         try
         {
             // Create Elasticsearch client
-            client = EsUtils.createClient(esUrl, authPath);
+            client = EsClientFactory.createRestClient(esUrl, authPath);
 
             // Create request
             Request req = new Request("POST", "/" + indexName + "/_update_by_query");
