+++ src/test/java/tt/TestRestClient.java
@@ -1,10 +1,7 @@
 package tt;
 
-import org.apache.http.HttpHost;
 import org.elasticsearch.client.RestClient;
-
-import gov.nasa.pds.registry.mgr.util.es.EsClientBuilder;
-import gov.nasa.pds.registry.mgr.util.es.EsUtils;
+import gov.nasa.pds.registry.mgr.es.client.EsClientFactory;
 
 
 public class TestRestClient
@@ -12,10 +9,7 @@ public class TestRestClient
 
     public static void main(String[] args) throws Exception
     {
-        HttpHost host = EsClientBuilder.parseUrl("my-host");
-        System.out.println(host);
-        
-        RestClient client = EsUtils.createClient("localhost", null);
+        RestClient client = EsClientFactory.createRestClient("localhost", null);
         client.close();
     }
